Day 3
This was a fun one, the first one that actually had me thinking a bit.
Part 1 is basically just a do-whatever problem, but the second part got
me thinking. I initially tried to do the dfs but with integers, this
turned out to be disfunctional, because a*10+b only works if you only
concat two integers, if there are more then it fails, because of spaces
in between, or so I think.
I went with a string-based dfs, that is also memoized, as I saw that
towards the end of a bank, you end up with recurring subproblems, thus
memoization works wonders here.