Day 5
Not too proud about this one. Part 1 is just a do-the-thing problem,
where you don't need to actually think, because the input allows for
some slow iterations. Regarding part 2, it works and I figured it out by
myself, but I'm pretty sure that this is much more hacky than it needs
to be. I might come back at some point to improve it. I see some peoples
solutions on reddit and mine is truly just awful. Just sharing this,
since it is truly my own solution.