Day 5 - Better Part 2
Inspired by
https://www.reddit.com/r/adventofcode/comments/1pemdwd/comment/nsdttqq/

I still wrote it a bit differently, in a way that it makes more sense to
me. I actually wrote this idea before settling with the thing from the
last commit, however I was trying to do this in place and access the
next element, rather than the previously inserted one. That caused a
bunch of annoying problems. I think this is the solution I will stick
with.