Day 6
This was quite fun. Quite unexpected part 2, it required me to throw
away all my parsing code and start over, but I realized that the
question is very easy once the text is transposed. You can basically
just transpose the text, split on \n\n and you have the numbers, then
you fish out the operator which ends up in the first line at the very
end and then you just use the numbers as in part 1. Code is bad, Go was
very annoying to work with here, functional languages (with sufficient
stdlibs) could've done the second part in a handful of lines probably.
Anyways quite a fun one, even if its mainly a parsing problem. I might
improve this code a bit when I come to it.

Shoutout to this thing
https://www.reddit.com/r/emacs/comments/1kpex3x/emacs_transpose_buffer/,
even if I'm not using Emacs