Day 8
This was by far the worst day so far. Coding Part 1 took a while until I
figured out the needed structure. Counting the moves inside a circuit is
quite odd and the overall data structures are weird.

I also got quite annoyed by Go, it lacks data structures so I planned to
use a library for sets for once, but the TreeSet would delete elements
that the CompareFunc returns 0 for, which is quite annoying because I'm
comparing floats.

Part 2 is not very efficient, but I don't feel like fighting with Go any
longer...