Day 10
Cool day...

Part 1 was pretty straightforward BFS

Part 2 however was fun, I got annoyed at Go again, but that seems to
become a regular occurrence. Initially I thought about using A*, but the
state space was too large. I knew that this was a Z3 problem, but
because all of the z3 bindings for Go are heavily outdated, I attempted
to go around it. I initially tried using another constraint solver
(centipede). It was bad, I regret it now. So I started implementing my
solution in Python Z3 which worked nicely. I wrote out the entire
solution on the REPL even. I later checked other peoples solutions and
saw one using Go, they were using a 3 year old Z3 binding which I
previously avoided, but I gave it a try anyways. It was quite hurtful
but here it is, the entire thing solved in Go Z3.

This day was quite fun, although I must say that Go was pretty annoying
to work with once again. The LSP died multiple times when I was adding a
libraries.