Day 11
Quite an easy day, if you know about memoized DFS that is, but we
already had this a few days back, so I don't think too many people had a
problem here.

I was quite surprised myself, that my solution worked that quickly,
especially since my answer on Part 1 was exactly 500, which sounded to
good to be true, but was actually the right answer.